requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.18.1
lxml==5.1.0
pandas==2.2.1
tqdm==4.66.2
fake-useragent==1.4.0
python-dotenv==1.0.1
aiohttp==3.9.3
asyncio==3.4.3
loguru==0.7.2
retry==0.9.2
langgraph==0.0.19
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.4
networkx==3.2.1
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1.post1
nltk==3.8.1
spacy==3.7.4
python-docx==1.1.0
openpyxl==3.1.2
flask==2.3.3
flask-wtf==1.2.1
jinja2==3.1.3
werkzeug==2.3.7
pdfkit==1.0.0
jieba==0.42.1
wordcloud==1.9.3
pyecharts==2.0.4
plotly==5.19.0 